Make turtlebot3_tests a canonical ros2 package

+++ ExternalTest/turtlebot3_tests/RunTB3WaypointFollowerTest.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 mkdir maps
-ln ../../Content/Turtlebot3_benchmark.pgm maps/Turtlebot3_benchmark.pgm
-ln ../../Content/Turtlebot3_benchmark.yaml maps/Turtlebot3_benchmark.yaml
-launch_test src/test_waypoint_follower.py tb3_model:=$1
+if [ ! -f maps/Turtlebot3_benchmark.pgm ]; then
+    ln ../../Content/Turtlebot3_benchmark.pgm maps/Turtlebot3_benchmark.pgm
+fi
+if [ ! -f maps/Turtlebot3_benchmark.yaml ]; then
+    ln ../../Content/Turtlebot3_benchmark.yaml maps/Turtlebot3_benchmark.yaml
+fi
+
+launch_test turtlebot3_tests/test_waypoint_follower.py tb3_model:=$1
